Removed redux, updated pathing convention, moved pages to folder, and isolated routes to own file
