+++ src/index.js
@@ -2,19 +2,14 @@ import React from 'react';
 import ReactDOM from 'react-dom';
 
 import createHistory from 'history/createBrowserHistory';
-import { Provider } from 'react-redux';
-import { ConnectedRouter } from 'react-router-redux';
-import { Route } from 'react-router';
-import store from './redux/store';
+import { Router } from 'react-router';
 import App from './App';
 
 const history = createHistory();
 
 ReactDOM.render(
-  <Provider store={store}>
-    <ConnectedRouter history={history}>
-      <App />
-    </ConnectedRouter>
-  </Provider>,
+  <Router history={history}>
+    <App />
+  </Router>,
   document.getElementById('root')
 );
