--- src/redux/actions/signupPageActions.js
@@ -1,4 +1,0 @@
-export const UPDATE_USERNAME = 'UPDATE_USERNAME';
-export const UPDATE_EMAIL = 'UPDATE_EMAIL';
-export const UPDATE_PASSWORD = 'UPDATE_PASSWORD';
-export const UPDATE_CONFIRM_PASSWORD = 'UPDATE_CONFIRM_PASSWORD';
