--- src/redux/reducers/index.js
@@ -1,8 +1,0 @@
-import { combineReducers } from 'redux';
-import { routerReducer } from 'react-router-redux';
-import SignupPageReducer from './signupPageReducer';
-
-export default combineReducers({
-  routing: routerReducer,
-  signupPage: SignupPageReducer,
-});
