--- src/redux/store.js
@@ -1,38 +1,0 @@
-import { createStore, applyMiddleware, compose } from 'redux';
-import { routerMiddleware } from 'react-router-redux';
-import thunk from 'redux-thunk';
-import createHistory from 'history/createBrowserHistory';
-import rootReducer from './reducers/';
-import { composeWithDevTools } from 'redux-devtools-extension';
-
-export const history = createHistory();
-
-const initialState = {
-  
-};
-const enhancers = [];
-const middleWare = [
-  thunk,
-  routerMiddleware(history),
-];
-
-if (process.env.NODE_ENV === 'development') {
-  const devToolsExtension = window.__REDUX_DEVTOOLS_EXTENSION__;
-
-  if (typeof devToolsExtension === 'function') {
-    enhancers.push(devToolsExtension());
-  }
-}
-
-const composedEnhancers = compose(
-  applyMiddleware(...middleWare),
-  ...enhancers
-);
-
-const store = createStore(
-  rootReducer,
-  initialState,
-  composedEnhancers
-);
-
-export default store;
